% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variational_bayes_EM.R
\name{.Estep2}
\alias{.Estep2}
\title{Expectation step for variational Bayesian inference in scGEM}
\usage{
.Estep2(
  d,
  Xid,
  Xcnt,
  n_topics,
  priors,
  g2,
  g3,
  g4,
  relation,
  subtree_size,
  max_est,
  likelihood_eps,
  count_eps,
  tree_mat,
  level_penalty
)
}
\description{
Expectation step for variational Bayesian inference in scGEM
}
