% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.getTreePrior}
\alias{.getTreePrior}
\title{Process the global tree and get global variables as prior}
\usage{
.getTreePrior(nHDP_tree, beta0, gamma1)
}
\arguments{
\item{nHDP_tree}{nHDP tree.}

\item{beta0}{the add-up when calculating the digamma function of count.}

\item{gamma1}{top level Dirichlet Process concentration}
}
\value{
list of global tree information
(ElnB: Elog(theta) (n_gene x n_topic)
 ElnPtop: ElnP (1 x n_topic))
}
\description{
Process the global tree and get global variables as prior
}
