% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.KMeans_L1}
\alias{.KMeans_L1}
\title{Mini-batch kmeans for dgCMatrix data, clusters that are less than a
required number of observations will be removed, the samples associated
with those clusters will be reassigned to the nearest cluster based on
L1 distance. Final clusters are sorted based on sample size.}
\usage{
.KMeans_L1(mat, K, iter_max = 3)
}
\arguments{
\item{mat}{dgCMatrix}

\item{K}{number of clusters}

\item{iter_max}{number of iteration (default is 3)}

\item{x}{dgCMatrix matrix}

\item{min_n}{minimum number of observations in each cluster}
}
\value{
list of two attributes: centers (p x K) and groups (1 x n)
}
\description{
Mini-batch kmeans for dgCMatrix data, clusters that are less than a
required number of observations will be removed, the samples associated
with those clusters will be reassigned to the nearest cluster based on
L1 distance. Final clusters are sorted based on sample size.
}
