% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.normData}
\alias{.normData}
\title{Normalize data as LogNormal in single cell (scale to 1e4 then take log)}
\usage{
.normData(x)
}
\arguments{
\item{x}{dgCMatrix format}
}
\value{
log-normalized data in dgCMatrix
}
\description{
Normalize data as LogNormal in single cell (scale to 1e4 then take log)
}
