% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variational_bayes_EM.R
\name{.updateCellWeight}
\alias{.updateCellWeight}
\title{Update the ElnP for cell}
\usage{
.updateCellWeight(
  cnt,
  all_parents,
  partition,
  picked_nodes,
  g2,
  g3,
  g4,
  tree_mat
)
}
\description{
Update the ElnP for cell
}
