% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGEM_tree.R
\name{.updateTree}
\alias{.updateTree}
\title{Update scGEM tree}
\usage{
.updateTree(
  x,
  scGEM_tree,
  model_paras,
  subtree_size,
  max_est,
  likelihood_eps,
  count_eps,
  learning_rate
)
}
\description{
Update scGEM tree
}
