% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGEM_tree.R
\name{initTree}
\alias{initTree}
\title{initialize the global tree based on nested mini-batch kmeans}
\usage{
initTree(
  x,
  num_topics,
  min_cells = NULL,
  blacklist_genes = NULL,
  rm.housekeeping_genes = TRUE
)
}
\arguments{
\item{x}{dgCMatrix of single cell count data (gene by cell).}

\item{num_topics}{vector of children topics per parent in each level.
(example: c(5, 4, 3))}

\item{min_cells}{minimum number of cells required for each topic,
topics fail to meet this requirement will be removed.
(default is min(num_topics))}

\item{blacklist_genes}{an expression of set of genes to be removed.
(default is NULL, example: "^MT|^RPS")}

\item{rm.housekeeping_genes}{whether or not to remove housekeeping genes.
(default is TRUE)}
}
\value{
centroids and counts of topics and parent-children relationship
}
\description{
initialize the global tree based on nested mini-batch kmeans
}
