% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGEM_tree.R
\name{minibatchInfer}
\alias{minibatchInfer}
\title{Infer scGEM tree based on mini-batch variational inference}
\usage{
minibatchInfer(
  X,
  scGEM_init,
  model_paras,
  batch_size = 1000,
  n_epoch = 20,
  subtree_size = c(1, 20),
  max_est = 50,
  likelihood_eps = 0.5 * 10^-3,
  count_eps = 0.5 * 10^-2,
  learning_rate = NULL,
  tol = 1e-04
)
}
\arguments{
\item{X}{dgCMatrix single cell data to learn}

\item{scGEM_init}{initialized global scGEM tree.}

\item{model_paras}{list of hyperparameters in the model.
b0: Dirichlet hyperparameter
g1: global Dirichlet process concentration
g2: local Dirichlet process concentration
g3: switching probability (stay)
g4: switching probability (searching down)}

\item{batch_size}{sample size for one mini-batch training.}

\item{n_epoch}{number of epochs (default is 20).}

\item{subtree_size}{size of the search tree for each cell.
(example: c(5, 20), at least 5, at most 20)}

\item{max_est}{maximum number of estimation for topic percentages in
each cell. (default is 25)}

\item{likelihood_eps}{stopping criteria for added likelihood in lower bound}

\item{count_eps}{stopping criteria for theta estimation in posterior}

\item{learning_rate}{learning rate of each batch (use ite ^ (-2) if NULL)}

\item{tol}{stopping criteria for learning, (LL(n)-LL(n-1))/LL(n-1)}
}
\value{
GEM distribution over cells, Gene distribution over GEM
}
\description{
Infer scGEM tree based on mini-batch variational inference
}
